Adding task manager function
These are one of the few advantages to having windows os on your computer. This os in my opinion is the most favorable while also having lots of versatility to the platform. Being able to upgrade my computer is a big concern when deciding on the computer I wanted to buy. I did not want to have to go to apple just to have a specialist with his special apple tool to crack open my computer or laptop when I can buy a windows computer and open that sucker with a screwdriver. Along with these windows os features you can find more when you buy a windows os device.